+++ CHANGELOG.md
@@ -16,6 +16,14 @@
 * Added `Hex::write_to_slice` method
 * Added `Hex::dot` method
 
+### Impls
+
+* Added `BitOr` implementations for `Hex`
+* Added `BitXor` implementations for `Hex`
+* Added `BitAnd` implementations for `Hex`
+* Added `Shr` implementations for `Hex`
+* Added `Shl` implementations for `Hex`
+
 ## 0.4.0 
 
 ### Directions